+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Loosened the `Show` instance of `Complex a` tp only depend on `Show a` (previously required `RealFloat a`)
 * Moved `showbRatioPrec` to `Text.Show.Text.Data.Ratio`, `showbComplexPrec` to `Text.Show.Text.Data.Complex`, `showbProxy` to `Text.Show.Text.Data.Proxy`, and `showbFingerprint` to `Text.Show.Text.GHC.Fingerprint`
 * Added `showbSpace` to `Text.Show.Text`
+* Added `mkShowList`, `mkShowListLazy`, and `mkShowbList` to `Text.Data.Text.TH`
 * Added the `Text.Show.Text.Data.OldTypeable` module for base >= 4.7
 * Added `Text.Show.Text.Debug.Trace` and `Text.Show.Text.Debug.Trace.TH`
 * Added the `Show1` class and corresponding instances for unary type constructors.